Rename helper file to bft
